requirements.txt
langchain==0.1.12
langchain-core==0.1.31
langgraph==0.0.26
anthropic==0.18.1
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.4
kubernetes==29.0.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-exporter-jaeger==1.23.0 